fix worker structure
